Move receipt formatting out of domain into presenter
