+++ src/main/java/com/ite/itea/application/usecase/OrderProductsUseCase.java
@@ -2,19 +2,22 @@ package com.ite.itea.application.usecase;
 
 import com.ite.itea.application.dto.OrderRequest;
 import com.ite.itea.application.dto.ReceiptDTO;
-import com.ite.itea.domain.CheckoutCalculator;
 import com.ite.itea.domain.retail.Order;
 import com.ite.itea.domain.retail.ProductId;
 import com.ite.itea.domain.retail.ProductRepository;
-import com.ite.itea.persistence.InMemoryProductRepository;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class OrderProductsUseCase {
 
-    private final ProductRepository productRepository = new InMemoryProductRepository();
-    private final CheckoutCalculator checkoutCalculator = new CheckoutCalculator();
+    private final ProductRepository productRepository;
+    private final ReceiptPresenter receiptPresenter;
+
+    public OrderProductsUseCase(ProductRepository productRepository, ReceiptPresenter receiptPresenter) {
+        this.productRepository = productRepository;
+        this.receiptPresenter = receiptPresenter;
+    }
 
     public ReceiptDTO execute(OrderRequest orderRequest) {
         final List<Order.OrderItem> orderItems = new ArrayList<>();
@@ -26,6 +29,10 @@ public class OrderProductsUseCase {
                     orderItems.add(new Order.OrderItem(value, itemRequest.amount())));
         }
 
-        return checkoutCalculator.prepareReceipt(new Order(orderItems));
+        return receiptPresenter.prepareReceipt(new Order(orderItems));
+    }
+
+    public interface ReceiptPresenter {
+        ReceiptDTO prepareReceipt(Order order);
     }
 }
