+++ src/main/java/com/ite/itea/endpoints/CheckoutController.java
@@ -4,6 +4,8 @@ import com.ite.itea.application.dto.OrderRequest;
 import com.ite.itea.application.dto.ReceiptDTO;
 import com.ite.itea.application.dto.ReceiptResponse;
 import com.ite.itea.application.usecase.OrderProductsUseCase;
+import com.ite.itea.persistence.InMemoryProductRepository;
+import com.ite.itea.presentation.ReceiptPresenter;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,7 +15,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @Controller
 public class CheckoutController {
 
-    private final OrderProductsUseCase orderProductsUseCase = new OrderProductsUseCase();
+    private final OrderProductsUseCase orderProductsUseCase = new OrderProductsUseCase(
+            new InMemoryProductRepository(),
+            new ReceiptPresenter()
+    );
 
     @ResponseBody
     @PostMapping(path = "/checkout",
